requirements.txt
discord.py >= 2.0.1
python-dotenv